requirements.txt
customtkinter
opencv-python
mediapipe
pillow
